// src/infrastructure/grpc/generated/course/types/section.ts
export interface UpdateSectionRequest {
  sectionId: string;
  userId: string;
  courseId: string;
  title: string;
  description: string;
  isPublished: boolean;
  order: number;
}

export interface DeleteSectionRequest {
  sectionId: string;
  courseId: string;
  userId: string;
}

export interface GetSectionsByCourseRequest {
  courseId: string;
}

export interface SectionData {
  id: string;
  courseId: string;
  title: string;
  lessons: LessonData[];
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
  description: string;
  isPublished: boolean;
  order: number;
  quiz: QuizData | undefined;
}

export interface SectionResponse {
  section?: SectionData | undefined;
  error?: Error | undefined;
}

export interface SectionsData {
  sections: SectionData[];
}

export interface SectionsResponse {
  sections?: SectionsData | undefined;
  error?: Error | undefined;
}

export interface DeleteSectionResponse {
  success?: DeleteSuccess | undefined;
  error?: Error | undefined;
}
// ...
export const GetSectionsByCourseRequest: MessageFns<GetSectionsByCourseRequest> = {
  encode(message: GetSectionsByCourseRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.courseId !== "") {
      writer.uint32(10).string(message.courseId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetSectionsByCourseRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetSectionsByCourseRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.courseId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetSectionsByCourseRequest {
    return { courseId: isSet(object.courseId) ? globalThis.String(object.courseId) : "" };
  },

  toJSON(message: GetSectionsByCourseRequest): unknown {
    const obj: any = {};
    if (message.courseId !== "") {
      obj.courseId = message.courseId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetSectionsByCourseRequest>, I>>(base?: I): GetSectionsByCourseRequest {
    return GetSectionsByCourseRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetSectionsByCourseRequest>, I>>(object: I): GetSectionsByCourseRequest {
    const message = createBaseGetSectionsByCourseRequest();
    message.courseId = object.courseId ?? "";
    return message;
  },
};
// ...
export const SectionResponse: MessageFns<SectionResponse> = {
  encode(message: SectionResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.section !== undefined) {
      SectionData.encode(message.section, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SectionResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSectionResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.section = SectionData.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SectionResponse {
    return {
      section: isSet(object.section) ? SectionData.fromJSON(object.section) : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: SectionResponse): unknown {
    const obj: any = {};
    if (message.section !== undefined) {
      obj.section = SectionData.toJSON(message.section);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SectionResponse>, I>>(base?: I): SectionResponse {
    return SectionResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SectionResponse>, I>>(object: I): SectionResponse {
    const message = createBaseSectionResponse();
    message.section = (object.section !== undefined && object.section !== null)
      ? SectionData.fromPartial(object.section)
      : undefined;
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    return message;
  },
};
// ...
export const SectionsResponse: MessageFns<SectionsResponse> = {
  encode(message: SectionsResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.sections !== undefined) {
      SectionsData.encode(message.sections, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SectionsResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSectionsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.sections = SectionsData.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SectionsResponse {
    return {
      sections: isSet(object.sections) ? SectionsData.fromJSON(object.sections) : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: SectionsResponse): unknown {
    const obj: any = {};
    if (message.sections !== undefined) {
      obj.sections = SectionsData.toJSON(message.sections);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SectionsResponse>, I>>(base?: I): SectionsResponse {
    return SectionsResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SectionsResponse>, I>>(object: I): SectionsResponse {
    const message = createBaseSectionsResponse();
    message.sections = (object.sections !== undefined && object.sections !== null)
      ? SectionsData.fromPartial(object.sections)
      : undefined;
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    return message;
  },
};

function createBaseDeleteSectionResponse(): DeleteSectionResponse {
  return { success: undefined, error: undefined };
}

export const DeleteSectionResponse: MessageFns<DeleteSectionResponse> = {
  encode(message: DeleteSectionResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.success !== undefined) {
      DeleteSuccess.encode(message.success, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DeleteSectionResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteSectionResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.success = DeleteSuccess.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DeleteSectionResponse {
    return {
      success: isSet(object.success) ? DeleteSuccess.fromJSON(object.success) : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: DeleteSectionResponse): unknown {
    const obj: any = {};
    if (message.success !== undefined) {
      obj.success = DeleteSuccess.toJSON(message.success);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DeleteSectionResponse>, I>>(base?: I): DeleteSectionResponse {
    return DeleteSectionResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DeleteSectionResponse>, I>>(object: I): DeleteSectionResponse {
    const message = createBaseDeleteSectionResponse();
    message.success = (object.success !== undefined && object.success !== null)
      ? DeleteSuccess.fromPartial(object.success)
      : undefined;
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}